gnome3.gnome-settings-daemon: bypass polkit by using an suid wrapper
This is intended to fix #14168 untli we have a better solution from
upstream GNOME.

The issue in brief is caused by the fact that gnome-settings-daemon is
non-restartable and if the policy definitions change in any way
it won't propagate to the running system because of that.

Also take note that we're abusing a debugging envar
`GSD_BACKLIGHT_HELPER` as it was never intended
for use outside testing.